Create index.html
As you see the demo or when you open download file you can see all source code which we used for this web page. This template is fully responsive with  html5 and css3 contents. Parallax Scrolling is a popular Effect and it’s easy to make. Here, we use css3 properties background fixed and Display table for make image perfectly center without any media query.